refactor(registry): split canonicalize tests
